requirements.txt
torch>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pypdf>=3.17.0
python-dotenv>=1.0.0
groq>=0.4.2
